GO-7111 Replace custom participant source with StaticSource and move init logic to participant.Init
Move detail setup, relation links, and template init from participantSource.ReadDoc
into participant.Init, matching the chat object pattern. Use StaticSource instead
of the custom participantSource.
